feat, ui, style : 스트레칭 모드 드롭다운 적용 및 합추기(리팩토링필요), 전체 css 및 코드정검
